Lot of Minor Bugs Fixed and Test Cases updated

+++ epan-components/xProduction/lib/Model/JobCard.php
@@ -43,7 +43,7 @@ class Model_JobCard extends \Model_Document{
 				->fieldQuery('name')
 			)->sortable(true);
 
-		$this->hasMany('xStore/Model_StockMovement','jobcard_id');
+		$this->hasMany('xStore/StockMovement','jobcard_id');
 		$this->hasMany('xProduction/JobCardAttachment','related_document_id',null,'Attachements');
 
 
@@ -52,7 +52,7 @@ class Model_JobCard extends \Model_Document{
 							// 'name|to_trim|required',
 							)
 					);
-		// $this->addHook('beforeDelete',$this);
+		$this->addHook('beforeDelete',$this);
 		// $this->add('dynamic_model/Controller_AutoCreator');
 	}
 
@@ -65,6 +65,8 @@ class Model_JobCard extends \Model_Document{
 	}	
 
 	function forceDelete(){
+		if(!$this->loaded()) throw new \Exception("Not Loaded", 1);
+		
 		$this->ref('xStore/StockMovement')->each(function($m){
 			$m->forceDelete();
 		});
